*: add Kubernetes tester

+++ internal/ec2/ec2.go
@@ -15,7 +15,6 @@ import (
 	"github.com/aws/aws-k8s-tester/internal/ssh"
 	"github.com/aws/aws-k8s-tester/pkg/awsapi"
 	"github.com/aws/aws-k8s-tester/pkg/zaputil"
-
 	"github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/aws-sdk-go/aws/session"
 	"github.com/aws/aws-sdk-go/service/cloudformation"
@@ -36,7 +35,7 @@ type Deployer interface {
 	Create() error
 	// Add creates one more instance to the cluster.
 	Add() error
-	// Stop stops create operation.å
+	// Stop stops create operation.
 	Stop()
 	// Delete deletes one instance.
 	Delete(id string) error
@@ -64,7 +63,7 @@ type embedded struct {
 	s3Buckets map[string]struct{}
 }
 
-// TODO: use cloudformation
+// TODO: use cloudformation, ASG
 
 // NewDeployer creates a new EKS deployer.
 func NewDeployer(cfg *ec2config.Config) (Deployer, error) {
@@ -96,7 +95,6 @@ func NewDeployer(cfg *ec2config.Config) (Deployer, error) {
 	if err != nil {
 		return nil, err
 	}
-
 	md.sts = sts.New(md.ss)
 	md.cf = cloudformation.New(md.ss)
 	md.ec2 = ec2.New(md.ss)
@@ -159,8 +157,10 @@ func (md *embedded) Create() (err error) {
 					md.lg.Warn("failed to revert VPC creation", zap.Error(derr))
 				}
 			}
-			if derr := md.deleteKeyPair(); derr != nil {
-				md.lg.Warn("failed to revert key pair creation", zap.Error(derr))
+			if md.cfg.KeyCreated {
+				if derr := md.deleteKeyPair(); derr != nil {
+					md.lg.Warn("failed to revert key pair creation", zap.Error(derr))
+				}
 			}
 		}
 	}()
@@ -169,6 +169,7 @@ func (md *embedded) Create() (err error) {
 	if err = catchStopc(md.lg, md.stopc, md.createKeyPair); err != nil {
 		return err
 	}
+
 	if md.cfg.VPCID != "" { // use existing VPC
 		if len(md.cfg.SubnetIDs) == 0 {
 			if err = catchStopc(md.lg, md.stopc, md.getSubnets); err != nil {
@@ -204,7 +205,7 @@ func (md *embedded) Create() (err error) {
 		} else {
 			md.cfg.VPCCIDR = *do.Vpcs[0].CidrBlock
 		}
-		md.lg.Info(
+		md.lg.Debug(
 			"found subnets",
 			zap.String("vpc-id", md.cfg.VPCID),
 			zap.String("vpc-cidr", md.cfg.VPCCIDR),
@@ -366,7 +367,7 @@ func (md *embedded) Delete(id string) (err error) {
 	}
 
 	now := time.Now().UTC()
-	md.lg.Info("deleting an instance", zap.String("cluster-name", md.cfg.ClusterName), zap.String("instance-id", id))
+	md.lg.Debug("deleting an instance", zap.String("cluster-name", md.cfg.ClusterName), zap.String("instance-id", id))
 
 	_, ok := md.cfg.Instances[id]
 	if !ok {
@@ -426,10 +427,15 @@ func (md *embedded) deleteInstances() (err error) {
 	for id := range md.cfg.Instances {
 		ids = append(ids, id)
 	}
+	if len(ids) == 0 {
+		md.lg.Warn("no EC2 instance found, nothing to delete", zap.String("cluster-name", md.cfg.ClusterName))
+		return nil
+	}
+
 	_, err = md.ec2.TerminateInstances(&ec2.TerminateInstancesInput{
 		InstanceIds: aws.StringSlice(ids),
 	})
-	md.lg.Info("terminating", zap.Strings("instance-ids", ids), zap.Error(err))
+	md.lg.Info("terminating", zap.String("cluster-name", md.cfg.ClusterName), zap.Strings("instance-ids", ids), zap.Error(err))
 
 	sleepDur := 5 * time.Second * time.Duration(md.cfg.ClusterSize)
 	if sleepDur > 3*time.Minute {
@@ -456,7 +462,7 @@ func (md *embedded) deleteInstances() (err error) {
 				}
 				if *inst.State.Name == "terminated" {
 					terminated[id] = struct{}{}
-					md.lg.Info("terminated", zap.String("instance-id", id))
+					md.lg.Info("terminated", zap.String("cluster-name", md.cfg.ClusterName), zap.String("instance-id", id))
 				}
 			}
 		}
@@ -464,6 +470,7 @@ func (md *embedded) deleteInstances() (err error) {
 	}
 
 	md.lg.Info("terminated",
+		zap.String("cluster-name", md.cfg.ClusterName),
 		zap.Strings("instance-ids", ids),
 		zap.String("request-started", humanize.RelTime(now, time.Now().UTC(), "ago", "from now")),
 	)
@@ -496,9 +503,11 @@ func (md *embedded) Terminate() (err error) {
 			errs = append(errs, err.Error())
 		}
 	}
-	if err = md.deleteKeyPair(); err != nil {
-		md.lg.Warn("failed to delete key pair", zap.Error(err))
-		errs = append(errs, err.Error())
+	if md.cfg.KeyCreated {
+		if err = md.deleteKeyPair(); err != nil {
+			md.lg.Warn("failed to delete key pair", zap.Error(err))
+			errs = append(errs, err.Error())
+		}
 	}
 
 	if len(errs) > 0 {
@@ -549,7 +558,7 @@ func (md *embedded) createInstances() (err error) {
 	// evenly distribute per subnet
 	left := md.cfg.ClusterSize
 
-	tokens := []string{}
+	tokens := make([]string, 0)
 	tknToCnt := make(map[string]int)
 
 	if md.cfg.ClusterSize > len(md.cfg.SubnetIDs) {
@@ -577,6 +586,10 @@ func (md *embedded) createInstances() (err error) {
 
 			if n < runInstancesBatch {
 				tkn := md.cfg.ClusterName + fmt.Sprintf("%X", time.Now().Nanosecond())
+				// otherwise, "InvalidParameterValue: Client token must be less than or equal to 64 characters"
+				if len(tkn) > 63 {
+					tkn = tkn[len(tkn)-63:]
+				}
 				tokens = append(tokens, tkn)
 
 				_, err = md.ec2.RunInstances(&ec2.RunInstancesInput{
@@ -611,6 +624,10 @@ func (md *embedded) createInstances() (err error) {
 				nLeft := n
 				for nLeft > 0 {
 					tkn := md.cfg.ClusterName + fmt.Sprintf("%X", time.Now().UTC().Nanosecond())
+					// otherwise, "InvalidParameterValue: Client token must be less than or equal to 64 characters"
+					if len(tkn) > 63 {
+						tkn = tkn[len(tkn)-63:]
+					}
 					tokens = append(tokens, tkn)
 
 					x := runInstancesBatch
@@ -659,6 +676,7 @@ func (md *embedded) createInstances() (err error) {
 
 			md.lg.Info(
 				"created EC2 instance group",
+				zap.String("cluster-name", md.cfg.ClusterName),
 				zap.String("subnet-id", subnetID),
 				zap.String("availability-zone", md.cfg.SubnetIDToAvailabilityZone[subnetID]),
 				zap.Int("instance-count", n),
@@ -668,6 +686,10 @@ func (md *embedded) createInstances() (err error) {
 		// create <1 instance per subnet
 		for i := 0; i < md.cfg.ClusterSize; i++ {
 			tkn := md.cfg.ClusterName + fmt.Sprintf("%X", time.Now().Nanosecond())
+			// otherwise, "InvalidParameterValue: Client token must be less than or equal to 64 characters"
+			if len(tkn) > 63 {
+				tkn = tkn[len(tkn)-63:]
+			}
 			tokens = append(tokens, tkn)
 			tknToCnt[tkn] = 1
 
@@ -706,6 +728,7 @@ func (md *embedded) createInstances() (err error) {
 
 			md.lg.Info(
 				"created EC2 instance group",
+				zap.String("cluster-name", md.cfg.ClusterName),
 				zap.String("subnet-id", subnetID),
 				zap.String("availability-zone", md.cfg.SubnetIDToAvailabilityZone[subnetID]),
 			)
@@ -749,6 +772,7 @@ func (md *embedded) createInstances() (err error) {
 							iv := ConvertEC2Instance(inst)
 							md.cfg.Instances[id] = iv
 							md.lg.Info("instance is ready",
+								zap.String("cluster-name", md.cfg.ClusterName),
 								zap.String("instance-id", iv.InstanceID),
 								zap.String("instance-public-ip", iv.PublicIP),
 								zap.String("instance-private-ip", iv.PrivateIP),
@@ -773,6 +797,7 @@ func (md *embedded) createInstances() (err error) {
 
 	md.lg.Info(
 		"created EC2 instances",
+		zap.String("cluster-name", md.cfg.ClusterName),
 		zap.Int("cluster-size", md.cfg.ClusterSize),
 		zap.String("request-started", humanize.RelTime(now, time.Now().UTC(), "ago", "from now")),
 	)
@@ -781,6 +806,7 @@ func (md *embedded) createInstances() (err error) {
 		md.cfg.Sync()
 		md.lg.Info(
 			"waiting for EC2 instances",
+			zap.String("cluster-name", md.cfg.ClusterName),
 			zap.Int("cluster-size", md.cfg.ClusterSize),
 			zap.String("request-started", humanize.RelTime(now, time.Now().UTC(), "ago", "from now")),
 		)
@@ -821,7 +847,7 @@ func (md *embedded) wait(mm map[string]ec2config.Instance) {
 				select {
 				case <-time.After(5 * time.Second):
 					out, err = sh.Run(
-						"tail -15 /var/log/cloud-init-output.log",
+						"tail -10 /var/log/cloud-init-output.log",
 						ssh.WithRetry(100, 5*time.Second),
 						ssh.WithTimeout(30*time.Second),
 					)
@@ -839,12 +865,12 @@ func (md *embedded) wait(mm map[string]ec2config.Instance) {
 
 					if IsReady(string(out)) {
 						sh.Close()
-						md.lg.Info("cloud-init-output.log READY!", zap.String("instance-id", id))
+						md.lg.Info("cloud-init-output.log READY!", zap.String("cluster-name", md.cfg.ClusterName), zap.String("instance-id", id))
 						delete(mm, id)
 						continue done
 					}
 
-					md.lg.Info("cloud-init-output NOT READY", zap.String("instance-id", id))
+					md.lg.Info("cloud-init-output NOT READY", zap.String("cluster-name", md.cfg.ClusterName), zap.String("instance-id", id))
 					fmt.Println(md.cfg.SSHCommands())
 				}
 			}
